// packages/react-native-payments-example/src/interface/request-options.interface.ts
export interface RequestOptionsInterface {
    asyncUpdate: boolean;
    coupon: boolean;
    requestBillingAddress: boolean;
    requestPayerEmail: boolean;
    requestPayerName: boolean;
    requestPayerPhone: boolean;
    requestShipping: boolean;
    showDisplayItems: boolean;
    totalValue: string;
}